requirements.txt
tensorflow==2.15.0
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.1
scikit-learn==1.3.2
matplotlib==3.8.2
requests==2.31.0